Fix RSS feed generation
The 'About' page was sometimes being added to the feed, whereas it
shouldn't. Followed the approach suggested by both the official docs[1]
and community forums[2], disabling feed generation for all page kinds
and enabling it just for the blog section.

Also took the opportunity to limit the number of items in the feed,
since by default it is unlimited.

[1]: https://gohugo.io/templates/rss/#configuration
[2]: https://discourse.gohugo.io/t/rss-feed-only-for-blog-posts/47558/2